+++ accel-pppd/extra/net-snmp/statPPTP.c
@@ -69,7 +69,7 @@ static int handle_statPPTPStarting(netsnmp_mib_handler *handler,
 				   netsnmp_agent_request_info *reqinfo,
 				   netsnmp_request_info *requests)
 {
-	unsigned int stat;
+	long stat;
 
 	(void)handler;
 	(void)reginfo;
@@ -93,7 +93,7 @@ static int handle_statPPTPActive(netsnmp_mib_handler *handler,
 				 netsnmp_agent_request_info *reqinfo,
 				 netsnmp_request_info *requests)
 {
-	unsigned int stat;
+	long stat;
 
 	(void)handler;
 	(void)reginfo;
